backend/requirements.txt

Flask==0.12.2
Flask-And-Redis==0.7
Flask-APScheduler
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.3.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-PyMongo==0.5.1
Flask-Redis==0.3.0
Flask-SocketIO==4.3.2
flatbuffers==1.12
h5py==2.10.0
matplotlib==3.3.2
numpy==1.19.3
opencv-python==4.2.0.34
pandas==1.2.3
Pillow==7.0.0
PyJWT==1.4.2
pymongo==3.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.14.2
python-socketio==4.6.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.4.1
webargs==7.0.1
Werkzeug==0.12.2
paho-mqtt==1.5.1